requirements.txt
openai==1.12.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pinecone-client==3.0.0
chromadb==0.4.22
psycopg2-binary==2.9.9
pgvector==0.2.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
tqdm==4.66.1
